+++ i18n/locales/en/common.json
@@ -317,6 +317,7 @@
   "Paste": "Paste",
   "Peers": "Peers",
   "Pending": "Pending",
+  "Pending multisignature transactions": "Pending multisignature transactions",
   "Pending...": "Pending...",
   "Percentage of successfully forged blocks in relation to all blocks (forged and missed).": "Percentage of successfully forged blocks in relation to all blocks (forged and missed).",
   "Percentage of total supply voting for a delegate.": "Percentage of total supply voting for a delegate.",
